pyproject.toml
[project]
name = "nanobabble"
version = "0.1.0"
description = "JAX/NNX trainer scaffold with checkpointing, metrics, and experiment tracking."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flax>=0.11.2",
    "jax>=0.7.1",
    "optax>=0.2.5",
    "orbax-checkpoint>=0.11.0",
    "polars>=1.30.0",
]
